Move objectUpdateReducer to objectReducers
