--- src/updateReducers.js
@@ -1,8 +1,0 @@
-
-export const objectUpdateReducer = (
-  state,
-  action
-) => ({
-  ...state,
-  ...action.data
-})
